requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
python-dotenv
openai
pydantic
aiofiles
